+++ backend/canisters/community/impl/src/updates/mod.rs
@@ -1,3 +1,4 @@
+pub mod accept_rules;
 pub mod add_members_to_channel;
 pub mod add_reaction;
 pub mod c2c_delete_community;