+++ backend/integration_tests/src/communities/send_message_tests.rs
@@ -83,6 +83,7 @@ fn send_crypto_in_channel() {
             sender_name: user1.username(),
             replies_to: None,
             mentioned: Vec::new(),
+            rules_accepted: None,
         },
     );